TINKERPOP-2285 Added ResponseError in gremlin-javascript
ResponseError in javascript is akin to ResponseException in java. It includes more direct access to protocol lever server error information such as the status code and status attributes which means we don't need to pack that information into the error message itself at all. Kept that old style to the error messages just in case users were parsing those in prior versions - didn't want to break that.